Pass config as string
