+++ client/src/cordova/apple/OutlineLib/VpnExtension/Sources/VpnExtension-Bridging-Header.h
@@ -1,0 +1,4 @@
+//
+//  Use this file to import your target's public headers that you would like to expose to Swift.
+//
+
