+++ client/src/cordova/apple/OutlineLib/VpnExtensionTest/VpnExtensionTest.swift
@@ -1,28 +1,17 @@
-// Copyright 2024 The Outline Authors
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+//  VpnExtensionTest.swift
+//  VpnExtensionTest
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//  Created by Vinicius Fortuna on 8/21/24.
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
 import XCTest
 
-import NetworkExtension
+@testable import VpnExtension
 
-@testable import OutlineTunnel
+final class VpnExtensionTest: XCTestCase {
 
-final class OutlineTunnelTest: XCTestCase {
-    
-    static let kAppGroup = "outline.spm.tests"
-    
-    func testSelectVpnAddress() {
+    func testSelectVpnAddress() throws {
         XCTAssertEqual("10.111.222.0", selectVpnAddress(interfaceAddresses:["172.16.9.2", "192.168.20.2", "169.254.19.1"]))
         XCTAssertEqual("172.16.9.1", selectVpnAddress(interfaceAddresses:["10.111.222.1", "192.168.20.2", "169.254.19.1"]))
         XCTAssertEqual("192.168.20.1", selectVpnAddress(interfaceAddresses:["10.111.222.1", "172.16.9.2", "169.254.19.1"]))
@@ -30,7 +19,7 @@ final class OutlineTunnelTest: XCTestCase {
         XCTAssertTrue(kVpnSubnetCandidates.values.contains(selectVpnAddress(interfaceAddresses: getNetworkInterfaceAddresses())))
     }
     
-    func testGetTunnelNetworkSettings() {
+    func testGetTunnelNetworkSettings() throws {
         let settings = OutlineTunnel.getTunnelNetworkSettings()
         
         XCTAssertEqual("::", settings.tunnelRemoteAddress)
